Add FasterWhisperHandler and utility functions for audio processing
